Fix for alerts in Chrome Browsers
This commit moves all the alert keyword in to a own _AlertKeywords
class. This did make it easier to fix the bug with WebDriverWait
class.
